migrate delegation confirmation
