+++ src/Staking/DelegationConfirmation/index.ts
@@ -1,2 +1,1 @@
 export * from './DelegationConfirmation'
-export * from './StakingCenterNavigator'
