+++ src/Staking/StakingCenterNavigator.tsx
@@ -3,11 +3,11 @@ import {createStackNavigator} from '@react-navigation/stack'
 import React from 'react'
 import {defineMessages, useIntl} from 'react-intl'
 
-import DelegationConfirmation from '../../legacy/components/Delegation/DelegationConfirmation'
 import {defaultNavigationOptions, defaultStackNavigatorOptions} from '../../legacy/navigationOptions'
 import {SEND_ROUTES, STAKING_CENTER_ROUTES, WALLET_ROOT_ROUTES} from '../../legacy/RoutesList'
 import {BiometricAuthScreen} from '../BiometricAuth'
 import {SettingsButton} from '../components/Button'
+import {DelegationConfirmation} from './DelegationConfirmation'
 import {StakingCenter} from './StakingCenter/StakingCenter'
 
 type StakingCenterRoutes = {
